Condensed samples to 16 bits

+++ src/boot/ksdk1.1.0/devMAX30105.h
@@ -7,4 +7,4 @@ CommStatus configureSensorMAX30105();
 
 CommStatus readSensorRegisterMAX30105(uint8_t deviceRegister, int numberOfBytes);
 
-CommStatus readNextSample(uint32_t * sample);
+CommStatus readNextSample(uint16_t *sample);
